Rename stats_socket to a new package dynconfig
